move PasswordAPI to separate file
